Issue-27 Added layers to the code
Broke the complete code into route, business and data layers,
for seperation of logic, and maintainability in future
