--- .env
@@ -1,9 +1,0 @@
-TIMES=2
-drive_private_key_id="8ff34ac90fe239ae46e28682af40d3bf26abea88"
-drive_private_key="-----BEGIN PRIVATE KEY-----\nMIIEvwIBADANBgkqhkiG9w0BAQEFAASCBKkwggSlAgEAAoIBAQCq16XGKLuDY/Uo\n9/+Bc+RfoP3pD8Cg4iX6TnrDtVT8FwcNcm8418hNV2/zioghHb/0R3bO598GR0W/\nAWmGiB6umRzrQ7bZukyY/S8R16+MPsuK5akDTJOFpwr3XsYv8/gVZQiO/qPJ7WkK\nD/XzAU4ScVHhJ1KMQY6IWnxmNYaNBrREHyDn7DJuQksgMdB6SNyTneumWH0/Ohcs\nOD37u+KaIHS+WgGVPpCxPZy9sXseSObCS0o8lKvMJMQVMMuA2+Bqv015cj+zVCQx\nnEKI7xr9bFTgGWqoA8J2/9LGUEH0E2e0DPpMjtZ+Szfk/dGyw7kaolTLcb9PRV+f\nxDSPmIfpAgMBAAECggEAEJSiVaORY1SZsBK1nQI66V8OFYb+Wn7oUa05M/JO37J6\ncD8pdTpYCc2RI3FP+erRvOsMqFMPraPQLJP+Al0GoUph6KFnRyVSFs0HuvMUageU\nKJRWfWf6r0+BfHWZxYCnlKf8Fyi2i7v7ZEsdQyrewWJ3IGN5Gvyn0Sr7BM5wExKp\nUItGvbO5DO1Zxhz+AXNYw26o+6lakZ80Ff7SJNS9ZqLZa9PKaoSmbP/+7jbtfIby\nTjq/qPqJBPlYk19vbXTaAzNWVA1bCD1YVL3aJ6mvIMFssbruK5ReS8x7wEImBCkq\nJ0FcWiSLPSgkpU3CuP+bVjJutJ4tH1dcyBXYhMswCwKBgQDfZX96TsdczE5lGX/B\nG7IKbCzmsgyUYdviZ5TdlN2pon1xUJDM2KtprgVSEqI9KBHFCEBsBMS1Y+8dFNRI\nOo54bceubkzYNc7U29d2CX0R0HINbULEbww668Eho+WfpMYGaR3YlyBsHPU1DPGt\npfO5upkLYINxNrQFIYucYYQUpwKBgQDDxqHFwFjn4RYaAigqCU8HffqIDO9jrwf3\nLhe/zfgSor4onHxXhNkliZekmETs4hgGxflS/Vf7aulhLoae4666ZQkbqDVHY2yx\n9dRMcmlUmogd01Q6dP7wr6iQOL0C7wp5fDe8SFeZdnic0PzTQtUtFcmJiAkFNJTP\nVhaZuVTA7wKBgQDWPL5OCzLB2eTsq9GK3gaWXDuLWgEaY+a9b1yiS1twjTYD76GN\n8dQhWxQbK2ePgtfXvxwY+1GyKw5rM0iJpC5xa8Ztj7X+O0wJcnNk7aCV7vMR0VCQ\nlXaDIwhqPIW/3gNSrohC6Fvdkr/NioZTY60i1sJi2MH1VuK5cPw0kohaTQKBgQCg\noLmccavAXNpwbnKbwtV3tTI4LDzB0hs8ADJ2Kb0vNR3GXlZazGIJvXGRF2n93aeO\n7yIYqEpmuxpOpQjJFtAGmN03NvS35OiKvsL4Eis/wiaKKhkCeXmu6sDny9WwsYYr\nlPJ55fURsbVWHNZDU6NqgWTBFb8o6TJrRxv19qyGOwKBgQCPgFVh277FfXO+/vJB\nt4o8Bk6qRM65/k958CkqQ262GnrsIuD+nxLVA06GV0aOv49xyRYsSMCQiSt4PLuj\n2HNwvwCR1ee5eqmErRSoSLeb6TUvs7w5DJAN8kjL8IitA5PCTIZh1Jvp2s/CI8EC\nxrEzPxPM5q24Ko3923PwrU2/pA==\n-----END PRIVATE KEY-----\n"
-drive_client_id="102686351296299622255"
-firebase_private_key_id="0f753c335071706d85d1b4fdee1ca9a7e7564964"
-firebase_private_key="-----BEGIN PRIVATE KEY-----\nMIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQCT+kNGFkZf/vKp\ncar84Ltiio5jChRA7Z72aNpKZInes20+WR3uA/UVCUWgh0R256Ut6IPrTYzwDa3o\nSxzPWJRGZZO5N3hVeKXGQracVLwChecbVfaIW9/wKvM63a0Bnl/IJo9dZJHOdJrA\nb5L7hpHcLUccAXXHIswFyg4sdba8x64IPjsWFc3UfP+v91bPHbEKHp0q2sNzY0V7\nQOBXJdhHoiY7EZtQ9UvuAdfCix17ctZyTHlA1VDr8FJFFBo1WM6RsLm3k7mK9FUs\n2IHh7y+NvNYkpj3qPyhbSBfjr5D1PtJDpioyl/C7ZeN+bF2xpkWTuvq30SuV6RHr\nZo2u4FefAgMBAAECggEAALprcLWXY+9lAOpPvJcHNQozD46nFXd3szpP7Cmb5HUQ\nHP8OQSjJX5KBaRQpyMle7Wuudwnzi9j8oYRou31QFXv/k7U2V+AM90hx22HYC7DL\nE25cnrrmrKHB7fwirdZSiDFfer9w8zCFQZsZcjOdf+AgCJJmLFYTe3FSEcfdnBoh\n/fFd8Ud3LOf0kZV8iMORnTqqsPcGmSeHa7eBoKsqSS5oPXv81S5H4gJtcGW0Rru/\nN1UqMgb/yI/4proMsvHDiqHCJ4UUA5YpSoFWZ3gq0Y9gyPaoD7I3eTW52zjxp5dn\nmqb4KgnsprXgUHPvWs1uO04YL0UO8YcZ5+IlrdevyQKBgQDGAoOnZt5aW/wUc64v\nuwjzYP14w5dXea9cnIWAFxxilGMVrXYxAhhhF2JfLKyNb5o68xCSwFJu1cModKa6\n0dINc7LgpXyUPHnyUhsRHjqflGrVV08XE8gEJHKL05d7wIweSADtl/xYkH7sCgXx\n7B8kDiYrk81HsgIskmC/vyhABQKBgQC/UKhpPHv4DPxLzDrUr75xg4Yt92aie6k9\nJQQ0bHVI2pwQ7H/kAuo2E5ZO/7eY6OL3b9W6bBfUKWRJ1bY94BD2/UxvxZ0Qj2Iw\n13w+PGZZX/sHKm5JSgvJiPV4YRU3CHRCSwwIt662HJXB3Fn+Pov8f8QXhPo16u+C\nCIKBP3ceUwKBgBiNK0HQcXWGvkTS6yAj8sIqnoMgzV0dS28SQUNWmkBm4bpIXcVJ\n2zf+rEXcxjBrTBZPGRj5G79+ZcpYXGwBpCfi1qUgH8hJj11KbOy8cMxknLKDNt1H\nqJ9K3aEUxU2k3DrM9sB+/d0FXZdqzr3mZx4TWwxAMU3YdoiMRc7Apv75AoGAFn+K\nvo1xWrNUPCmxRW/BaqH39pHgjP0T4u/mda/HfyDON6NISu0Y0/ErvPCPi7jfcA/D\nx5+MkdOoSpKKUu51sEYLZh2MIfVqq6m7n6oEsz/5SldwDshZaILlsxlfoFArM9yV\ndrT7I8lqxxh9sMt0NOZljgqCTjLppZYQmth2zqcCgYB7EgYIEyBxpylhDSFOxx/Y\nLzJ4aai8I8E9l1Hf3qopk7BZ8K3NpzoNPBn/eYVu+98YdfZbs+nophMsUxDts0A/\nTAjSFeYIPgkeAOmmu0w13z/hXdXe4iShvb4xpuT6PzUJpgaSaMH18hnFO5p5zyyx\nI86hyyH1vqFFMdsy4IhudA==\n-----END PRIVATE KEY-----\n"
-firebase_client_id="109106132237287005941"
-littlehelp_key="4JAN2017"
-heroku_db_uri="postgres://rsjworyabnzayq:9d58468df927adbbb5e0b61178ed465f89bdee5f80b8854d882dc7c3517a7c4f@ec2-174-129-254-218.compute-1.amazonaws.com:5432/da3rkhlknkfbsh"
